requirements.txt
absl-py==0.9.0
aiofiles==0.4.0
alabaster==0.7.12
alembic==1.2.0
apache-airflow==1.10.10
apispec==1.3.3
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.11.1
asn1crypto==0.24.0
astroid==2.4.0
astropy==4.0.1.post1
astunparse==1.6.3
attrs==19.3.0
Automat==0.7.0
awscli==1.18.50
Babel==2.8.0
backcall==0.1.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bleach==2.1.4
blis==0.4.1
boto==2.49.0
boto3==1.10.33
botocore==1.16.0
cached-property==1.5.1
cachetools==4.1.0
catalogue==1.0.0
catboost==0.20.2
cattrs==0.9.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
chart-studio==1.1.0
Click==7.0
cloudpickle==1.4.1
colorama==0.4.3
colorlog==4.0.2
colorlover==0.3.0
combo==0.1.0
configparser==3.5.3
confuse==1.1.0
constantly==15.1.0
cos-utils==0.0.13
croniter==0.3.32
cryptography==2.3.1
cufflinks==0.17.0
cvxpy==1.0.31
cycler==0.10.0
cymem==2.0.3
cypari==2.4.0
dash==1.3.1
dash-core-components==1.2.1
dash-html-components==1.0.1
dash-renderer==1.1.0
dash-table==4.3.0
dask==2.18.1
datefinder==0.7.0
DateTime==4.3
decorator==4.3.0
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
django-os-geocoder==0.1.6
dnspython==1.16.0
docutils==0.15.2
ecos==2.0.7.post1
email-validator==1.1.1
entrypoints==0.2.3
fancyimpute==0.5.4
fb-idb==0.0.7
filelock==3.0.12
findspark==1.3.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-Compress==1.4.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.2
funcy==1.14
future==0.17.1
FXrays==1.3.3
gast==0.3.3
gensim==3.8.2
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14
grpcio==1.28.1
grpclib==0.3.0
gunicorn==19.9.0
h2==3.1.1
h5py==2.10.0
hpack==3.0.0
html5lib==1.0.1
htmlmin==0.1.12
hyperframe==5.2.0
hyperlink==18.0.0
hyperopt==0.2.4
ibm-cos-sdk==2.6.2
ibm-cos-sdk-core==2.6.2
ibm-cos-sdk-s3transfer==2.6.2
ibmos2spark==1.0.1
idna==2.8
imagesize==1.2.0
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jedi==0.12.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kmodes==0.10.1
knnimpute==0.1.0
lazy-object-proxy==1.4.3
lightgbm==2.3.1
llvmlite==0.32.0
lockfile==0.12.2
logger==1.4
luigi==2.8.9
lxml==4.4.1
lyricsgenius==1.3.1
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.2.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.3
mlxtend==0.17.2
more-itertools==8.2.0
multidict==4.5.2
multiprocess==0.70.9
murmurhash==1.0.2
mxnet==1.6.0
natsort==7.0.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.4
nltk==3.5
notebook==5.7.8
numba==0.49.0
numexpr==2.7.1
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
osqp==0.6.1
packaging==20.3
pandas==1.0.3
pandas-datareader==0.8.0
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pbr==5.4.3
pendulum==1.4.4
pexpect==4.8.0
phik==0.9.11
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
plink==2.3
plotly==4.4.1
pluggy==0.13.1
preshed==3.0.2
prison==0.1.3
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==1.0.18
psutil==5.7.0
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==1.0.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyLDAvis==2.1.2
pylint==2.5.0
pymaybe==0.1.6
PyMySQL==0.9.3
pyod==0.7.8.2
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.15.4
PySocks==1.7.1
pyspark==2.4.5
pyspark-stubs==2.4.0.post7
pytest==5.4.1
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==19.0.0
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.4.4
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.3.3
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.4.1
scs==2.1.2
seaborn==0.10.1
Send2Trash==1.5.0
setproctitle==1.1.10
shap==0.32.1
simplegeneric==0.8.1
six==1.14.0
sklearn==0.0
smart-open==2.0.0
snappy==2.7
snappy-manifolds==1.1.1
snowballstemmer==2.0.0
South==1.0.2
spacy==2.2.4
spherogram==1.8.2
Sphinx==3.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.11
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
srsly==1.0.2
stevedore==1.31.0
suod==0.0.4
svgwrite==1.4
tabulate==0.8.7
tbb==2019.0
tenacity==4.12.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
textblob==0.15.3
thinc==7.4.0
threadpoolctl==2.0.0
thrift==0.13.0
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.40.0
traitlets==4.3.2
Tree==0.2.4
tweepy==3.8.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
umap-learn==0.4.2
unicodecsv==0.14.1
urllib3==1.25.7
versioneer==0.18
virtualenv==20.0.18
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.4.2
wordcloud==1.6.0
wrapt==1.12.1
WTForms==2.2.1
xgboost==0.90
yellowbrick==1.0.1
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.1.0
